+++ console/headers.go
@@ -50,18 +50,11 @@ type ColumnDefinitions []ColumnDefinition
 func (d ColumnDefinitions) resolveKey(name string) string {
 	low := strings.ToLower(name)
 	for _, def := range d {
-		if low == strings.ToLower(def.propKey()) {
+		switch low {
+		case strings.ToLower(def.propKey()), strings.ToLower(def.title(false)):
 			return def.propKey()
 		}
-		switch def.(type) {
-		case StringColumnDefinition:
-			sdef := def.(StringColumnDefinition)
-			if low == strings.ToLower(sdef.Friendly) {
-				return def.propKey()
-			}
-		}
 	}
-
 	return ""
 }