+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bugfixes
 
 - Fix regression: run templates/one-liners failed on `storageobject`, `subscription` entities
+- Filtering in `awless list --filter` now works with column types other than string
 
 ## v0.0.20 [2017-03-20]